requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
billiard==4.2.4
blis==1.3.3
catalogue==2.0.10
celery==5.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
confection==1.3.3
cymem==2.0.13
frozenlist==1.8.0
gigachat==0.2.0
h11==0.16.0
hashids==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jericho==3.3.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain-core==1.3.2
langchain-gigachat==0.5.0
langchain-protocol==0.0.12
langsmith==0.7.36
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mementos==1.3.1
more-itertools==11.0.2
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==26.2
preshed==3.0.13
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.7.4
pydantic-settings==2.14.0
pydantic_core==2.18.4
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==6.4.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
TatSu==5.8.3
tenacity==9.1.4
termcolor==3.3.0
textworld==1.7.0
thinc==8.3.13
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
